move getsizeof_total_<type> tests to single file
